requirements.txt
telethon
python-dotenv
pandas
requests
tqdm
psycopg2-binary
dbt
dagster
python-dotenv
dbt-postgres
# psycopg2-binary is used for PostgreSQL database connections